Updates for building under RISC OS with a C89 compiler.
The build under RISC OS is not especially involved - it is largely
the result of converting the code to build with my earlier compiler.
The result is that it will build with the cross compiling environment
and produces the exact same output. In theory this can be used to
produce SVGs from bitmaps, when those bitmaps are simple block
diagrams.